+++ src/Nethermind/Nethermind.Consensus/Processing/PrewarmerTxAdapter.cs
@@ -1,0 +1,25 @@
+// SPDX-FileCopyrightText: 2025 Demerzel Solutions Limited
+// SPDX-License-Identifier: LGPL-3.0-only
+
+using Nethermind.Core;
+using Nethermind.Evm;
+using Nethermind.Evm.State;
+using Nethermind.Evm.Tracing;
+using Nethermind.Evm.TransactionProcessing;
+using Nethermind.State;
+
+namespace Nethermind.Consensus.Processing;
+
+public class PrewarmerTxAdapter(ITransactionProcessorAdapter baseAdapter, BlockCachePreWarmer preWarmer, IWorldState worldState) : ITransactionProcessorAdapter
+{
+    public TransactionResult Execute(Transaction transaction, ITxTracer txTracer)
+    {
+        if (worldState is IPreBlockCaches preBlockCaches && preBlockCaches.IsWarmWorldState)
+        {
+            preWarmer.OnBeforeTxExecution(transaction);
+        }
+        return baseAdapter.Execute(transaction, txTracer);
+    }
+
+    public void SetBlockExecutionContext(in BlockExecutionContext blockExecutionContext) => baseAdapter.SetBlockExecutionContext(in blockExecutionContext);
+}
