+++ src/Nethermind/Nethermind.State/IPreBlockCaches.cs
@@ -6,5 +6,5 @@ namespace Nethermind.State;
 public interface IPreBlockCaches
 {
     PreBlockCaches Caches { get; }
-    bool IsToBeWarmedWorldState { get; }
+    bool IsWarmWorldState { get; }
 }
